requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pafy==0.5.5
python-vlc==3.0.11115
requests==2.25.0
six==1.15.0
unlimited-youtube-search==0.0.2
urllib3==1.26.2
Werkzeug==1.0.1
youtube-dl==2020.11.21.1
